[Tactics] 3d map with blocks

+++ Tactics/Engine/include/UI/WindowMessages.h
@@ -22,9 +22,9 @@ struct KeyPressed : Message
 
 struct ClickOn : Message
 {
-    ClickOn(const class Object* obj, std::uint16_t sub) : object(obj), sub(sub) {}
+    ClickOn(const class Object* obj, std::uint32_t sub) : object(obj), sub(sub) {}
     const Object* object;
-    std::uint16_t sub;
+    std::uint32_t sub;
 };
 
 struct ScrollWheel : Message
